Relocate classes and rename artifact groupId
