+++ pom.xml
@@ -3,9 +3,9 @@
 
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.expediagroup.dropwizard.bundle</groupId>
+    <groupId>com.expediagroup</groupId>
     <artifactId>dropwizard-resilience4j-bundle</artifactId>
-    <version>1.0.1-SNAPSHOT</version>
+    <version>2.0.0-SNAPSHOT</version>
     <packaging>jar</packaging>
 
     <name>dropwizard-resilience4j-bundle</name>
