+++ src/test/java/com/expediagroup/dropwizard/resilience4j/ResourceTestUtil.java
@@ -1,4 +1,4 @@
-package com.homeaway.dropwizard.bundle.resilience4j;
+package com.expediagroup.dropwizard.resilience4j;
 
 import java.io.File;
 import java.net.URI;
